+++ db/silkworm/db/access_layer.hpp
@@ -24,37 +24,11 @@
 #include <vector>
 
 #include <silkworm/chain/config.hpp>
-#include <silkworm/common/magic_enum.hpp>
+#include <silkworm/common/rlp_err.hpp>
 #include <silkworm/db/mdbx.hpp>
 #include <silkworm/db/util.hpp>
 #include <silkworm/types/account.hpp>
 #include <silkworm/types/block.hpp>
-
-namespace silkworm::rlp {
-
-class DecodingError : public std::exception {
-  public:
-    explicit DecodingError(DecodingResult err)
-        : err_{magic_enum::enum_integer<DecodingResult>(err)},
-          message_{"Decoding error : " + std::string(magic_enum::enum_name<DecodingResult>(err))} {};
-    explicit DecodingError(DecodingResult err, const std::string& message)
-        : err_{magic_enum::enum_integer<DecodingResult>(err)}, message_{message} {};
-    virtual ~DecodingError() noexcept {};
-    const char* what() const noexcept override { return message_.c_str(); }
-    int err() const noexcept { return err_; }
-
-  protected:
-    int err_;
-    std::string message_;
-};
-
-inline void err_handler(DecodingResult err) {
-    if (err != DecodingResult::kOk) {
-        throw DecodingError(err);
-    }
-}
-
-}  // namespace silkworm::rlp
 
 namespace silkworm::db {
